Refactored wav file recording code into seperate file and fixed the progress bar for audio recording.
