+++ .idea/modules.xml
@@ -4,6 +4,7 @@
     <modules>
       <module fileurl="file://$PROJECT_DIR$/DigitalStethoscope.iml" filepath="$PROJECT_DIR$/DigitalStethoscope.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/musicg_android_demo/musicg_android_demo.iml" filepath="$PROJECT_DIR$/musicg_android_demo/musicg_android_demo.iml" />
     </modules>
   </component>
 </project>
