+++ app/src/main/java/com/example/pavan/digitalstethoscope/MainActivity.java
@@ -358,28 +358,6 @@ public class MainActivity extends AppCompatActivity {
             @Override
             public void onClick(View v) {
 
-                // TODO: audio record code here
-                recorder = new AudioRecord(MediaRecorder.AudioSource.MIC,
-                        RECORDER_SAMPLERATE, RECORDER_CHANNELS,RECORDER_AUDIO_ENCODING, bufferSize);
-
-                int i = recorder.getState();
-                if(i==1)
-                    recorder.startRecording();
-                isRecording = true;
-                recordingThread = new Thread(new Runnable() {
-
-                    @Override
-                    public void run() {
-                        writeAudioDataToFile();
-                    }
-                },"AudioRecorder Thread");
-                recordingThread.start();
-                long startTime = System.currentTimeMillis();
-                long time = 0;
-                while (time != (startTime + 10000))
-                    time = System.currentTimeMillis();
-                stopRecording();
-
 
 
                 //pg bar thread runs for 10 sec.
@@ -388,6 +366,8 @@ public class MainActivity extends AppCompatActivity {
                     @Override
                     public void run()
                     {
+                        WavRecorder wavRecorder = new WavRecorder("/storage/self/primary/Audio/example.wav");
+                        wavRecorder.startRecording();
                         while(recordPgBarStatus<=100)
                         {
 
@@ -403,173 +383,13 @@ public class MainActivity extends AppCompatActivity {
                             });
                             recordPgBarStatus+=10;
                         }
+                        wavRecorder.stopRecording();
                     }
                 }).start();
-            }
-            private void stopRecording() {
-                if(null != recorder){
-                    isRecording = false;
 
-                    int i = recorder.getState();
-                    if(i==1) {
-                        recorder.stop();
-                    }
-                    recorder.release();
-
-                    recorder = null;
-                    recordingThread = null;
-
-                }
-
-                copyWaveFile(getTempFilename(),getFilename());
-                deleteTempFile();
-            }
-
-            private void copyWaveFile(String tempFilename, Object filename) {
-                FileInputStream in;
-                FileOutputStream out;
-                long totalAudioLen = 0;
-                long totalDataLen = totalAudioLen + 36;
-                int channels = 2;
-                long byteRate = RECORDER_BPP * RECORDER_SAMPLERATE * channels/8;
-
-                byte[] data = new byte[bufferSize];
-
-                try {
-                    in = new FileInputStream(tempFilename);
-                    out = new FileOutputStream(String.valueOf(filename));
-                    totalAudioLen = in.getChannel().size();
-                    totalDataLen = totalAudioLen + 36;
-                    WriteWaveFileHeader(out, totalAudioLen, totalDataLen,
-                            (long) RECORDER_SAMPLERATE, channels, byteRate);
-
-                    while(in.read(data) != -1){
-                        out.write(data);
-                    }
-
-                    in.close();
-                    out.close();
-                } catch (FileNotFoundException e) {
-                    e.printStackTrace();
-                } catch (IOException e) {
-                    e.printStackTrace();
-                }
-            }
-
-            private void WriteWaveFileHeader(
-                    FileOutputStream out, long totalAudioLen,
-                    long totalDataLen, long recorderSamplerate, int channels,
-                    long byteRate)throws IOException {
-                byte[] header = new byte[44];
-
-                header[0] = 'R'; // RIFF/WAVE header
-                header[1] = 'I';
-                header[2] = 'F';
-                header[3] = 'F';
-                header[4] = (byte) (totalDataLen & 0xff);
-                header[5] = (byte) ((totalDataLen >> 8) & 0xff);
-                header[6] = (byte) ((totalDataLen >> 16) & 0xff);
-                header[7] = (byte) ((totalDataLen >> 24) & 0xff);
-                header[8] = 'W';
-                header[9] = 'A';
-                header[10] = 'V';
-                header[11] = 'E';
-                header[12] = 'f'; // 'fmt ' chunk
-                header[13] = 'm';
-                header[14] = 't';
-                header[15] = ' ';
-                header[16] = 16; // 4 bytes: size of 'fmt ' chunk
-                header[17] = 0;
-                header[18] = 0;
-                header[19] = 0;
-                header[20] = 1; // format = 1
-                header[21] = 0;
-                header[22] = (byte) channels;
-                header[23] = 0;
-                header[24] = (byte) (recorderSamplerate & 0xff);
-                header[25] = (byte) ((recorderSamplerate >> 8) & 0xff);
-                header[26] = (byte) ((recorderSamplerate >> 16) & 0xff);
-                header[27] = (byte) ((recorderSamplerate >> 24) & 0xff);
-                header[28] = (byte) (byteRate & 0xff);
-                header[29] = (byte) ((byteRate >> 8) & 0xff);
-                header[30] = (byte) ((byteRate >> 16) & 0xff);
-                header[31] = (byte) ((byteRate >> 24) & 0xff);
-                header[32] = (byte) (2 * 16 / 8); // block align
-                header[33] = 0;
-                header[34] = RECORDER_BPP; // bits per sample
-                header[35] = 0;
-                header[36] = 'd';
-                header[37] = 'a';
-                header[38] = 't';
-                header[39] = 'a';
-                header[40] = (byte) (totalAudioLen & 0xff);
-                header[41] = (byte) ((totalAudioLen >> 8) & 0xff);
-                header[42] = (byte) ((totalAudioLen >> 16) & 0xff);
-                header[43] = (byte) ((totalAudioLen >> 24) & 0xff);
-
-                out.write(header, 0, 44);
-            }
-
-            private void deleteTempFile() {
-                File file = new File(getTempFilename());
-
-                file.delete();
-            }
-
-            private Object getFilename() {
-                String filepath = "/storage/emulated/0/Audio"+"/audiorecordtest.wav";
-                File file = new File(filepath);
-
-
-                return (file.getAbsolutePath());
 
             }
 
-            private void writeAudioDataToFile() {
-                byte data[] = new byte[bufferSize];
-                String filename = getTempFilename();
-                FileOutputStream os = null;
-
-                try {
-                    os = new FileOutputStream(filename);
-                } catch (FileNotFoundException e) {
-
-                    e.printStackTrace();
-                }
-
-                int read;
-
-                if(null != os){
-                    while(isRecording){
-                        read = recorder.read(data, 0, bufferSize);
-
-                        if(AudioRecord.ERROR_INVALID_OPERATION != read){
-                            try {
-                                os.write(data);
-                            } catch (IOException e) {
-                                e.printStackTrace();
-                            }
-                        }
-                    }
-
-                    try {
-                        os.close();
-                    } catch (IOException e) {
-                        e.printStackTrace();
-                    }
-                }
-            }
-
-            private String getTempFilename() {
-                String filepath = "/storage/emulated/0/Audio"+"/audiorecordtest2.wav";
-
-
-                File tempFile = new File(filepath);
-
-
-
-                return (tempFile.getAbsolutePath());
-            }
         });
